+++ examples/dataset_examples.py
@@ -3,8 +3,12 @@ from wildboar.datasets import (
     list_datasets,
     list_repositories,
     list_bundles,
+    clear_cache,
+    set_cache_dir,
 )
 
+set_cache_dir("wildboar_dataset_cache")
+clear_cache()
 x, y = load_dataset("Wafer", repository="wildboar/ucr")
 print(list_datasets(repository="wildboar/ucr-tiny"))
 print(list_repositories())